Refactor API client, add icons and navigation updates

+++ src/api/auth.ts
@@ -1,13 +1,11 @@
-import { ApiUrl } from '@configs/apiUrl';
-import { client } from '@configs/axiosConfig';
-import { extraParams } from '@utils/apiResponse';
-import { AppApiTypeRequest } from '@api/typeRequest';
+import { ApiUrl } from '@/constants/api';
+import { client } from '@/api/apiClient';
+import { extraParams } from '@/utils/apiResponse';
+import { AppApiTypeRequest } from '@/api/typeRequest';
 
 export class AuthApi {
   static async getUserInfo(params: any) {
-    const response = await client.get(
-      ApiUrl.auth.get_user_info + extraParams(params),
-    );
+    const response = await client.get(ApiUrl.auth.get_user_info + extraParams(params));
     return response.data;
   }
   static async LoginFacebook(params: AppApiTypeRequest.LoginFacebook) {
@@ -18,17 +16,20 @@ export class AuthApi {
     const response = await client.post(ApiUrl.auth.login_google, params);
     return response;
   }
+  static async LoginZalo(params: AppApiTypeRequest.LoginGoogle) {
+    const response = await client.post(ApiUrl.auth.login_zalo, params);
+    return response;
+  }
+  static async LoginApple(params: AppApiTypeRequest.LoginGoogle) {
+    const response = await client.post(ApiUrl.auth.login_apple, params);
+    return response;
+  }
   static async DeletedAccount(params: any) {
-    const response = await client.delete(
-      `${ApiUrl.auth.delete_account}/${params}`,
-    );
+    const response = await client.delete(`${ApiUrl.auth.delete_account}/${params}`);
     return response;
   }
   static async UpdateUserInfo(_id: string, params: any) {
-    const response = await client.patch(
-      `${ApiUrl.auth.update_user_info}/${_id}`,
-      params,
-    );
+    const response = await client.patch(`${ApiUrl.auth.update_user_info}/${_id}`, params);
     return response;
   }
 }
